+++ db/migrate/20221026162057_add_ispf_theme_to_activities.rb
@@ -1,0 +1,5 @@
+class AddIspfThemeToActivities < ActiveRecord::Migration[6.1]
+  def change
+    add_column :activities, :ispf_theme, :integer
+  end
+end