+++ CHANGELOG.md
@@ -1122,6 +1122,16 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 
 ## [unreleased]
 
+- Update seeds and create data migration for adding ISPF fund entity to the service
+- Hide ISPF reports when the ISPF feature flag is enabled
+- ISPF programmes cannot be created when the feature flag is enabled
+- ISPF activities are not shown to users when the feature flag is enabled
+- ISPF activities are not shown in search results when the feature flag is enabled
+- Exclude ISPF-related entities from exports when the feature flag is enabled
+- The first step for an ISPF programme is now choosing between ODA and Non-ODA
+- Add ISPF theme form step
+- Add ISPF partner countries form step
+
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-121...HEAD
 [release-121]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-120...release-121
 [release-120]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-119...release-120